+++ src/index.ts
@@ -10,3 +10,4 @@ export * from "./logger.js";
 export * from "./errors.js";
 export * from "./cache-status.js";
 export * from "./target-resolver.js";
+export * from "./deps.js";